+++ app/settings.py
@@ -7,7 +7,7 @@ import yaml
 
 
 
-VERSION = "v0.8.1"
+VERSION = "v0.8.2"
 
 # --- Environment Configuration ---
 BASE_DIR = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))